backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
mammoth
supabase==2.0.3
google-generativeai
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
